@@ -15,7 +15,7 @@ import org.jooq.impl.CatalogImpl;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class DefaultCatalog extends CatalogImpl {
 
     private static final long serialVersionUID = 1L;
@@ -26,9 +26,9 @@ public class DefaultCatalog extends CatalogImpl {
     public static final DefaultCatalog DEFAULT_CATALOG = new DefaultCatalog();
 
     /**
-     * The schema <code>minance</code>.
+     * The schema <code>DEFAULT_SCHEMA</code>.
      */
-    public final Minance MINANCE = Minance.MINANCE;
+    public final DefaultSchema DEFAULT_SCHEMA = DefaultSchema.DEFAULT_SCHEMA;
 
     /**
      * No further instances allowed
@@ -40,15 +40,15 @@ public class DefaultCatalog extends CatalogImpl {
     @Override
     public final List<Schema> getSchemas() {
         return Arrays.asList(
-            Minance.MINANCE
+            DefaultSchema.DEFAULT_SCHEMA
         );
     }
 
     /**
-     * A reference to the 3.18 minor release of the code generator. If this
+     * A reference to the 3.19 minor release of the code generator. If this
      * doesn't compile, it's because the runtime library uses an older minor
-     * release, namely: 3.18. You can turn off the generation of this reference
+     * release, namely: 3.19. You can turn off the generation of this reference
      * by specifying /configuration/generator/generate/jooqVersionReference
      */
-    private static final String REQUIRE_RUNTIME_JOOQ_VERSION = Constants.VERSION_3_18;
+    private static final String REQUIRE_RUNTIME_JOOQ_VERSION = Constants.VERSION_3_19;
 }
