@@ -13,7 +13,6 @@ import org.jooq.impl.SchemaImpl;
 
 import today.ihelio.jooq.tables.Accounts;
 import today.ihelio.jooq.tables.Banks;
-import today.ihelio.jooq.tables.FlywaySchemaHistory;
 import today.ihelio.jooq.tables.MinanceCategory;
 import today.ihelio.jooq.tables.RawCategoryToMinanceCategory;
 import today.ihelio.jooq.tables.Transactions;
@@ -22,51 +21,46 @@ import today.ihelio.jooq.tables.Transactions;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Minance extends SchemaImpl {
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class DefaultSchema extends SchemaImpl {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * The reference instance of <code>minance</code>
+     * The reference instance of <code>DEFAULT_SCHEMA</code>
      */
-    public static final Minance MINANCE = new Minance();
+    public static final DefaultSchema DEFAULT_SCHEMA = new DefaultSchema();
 
     /**
-     * The table <code>minance.accounts</code>.
+     * The table <code>accounts</code>.
      */
     public final Accounts ACCOUNTS = Accounts.ACCOUNTS;
 
     /**
-     * The table <code>minance.banks</code>.
+     * The table <code>banks</code>.
      */
     public final Banks BANKS = Banks.BANKS;
 
     /**
-     * The table <code>minance.flyway_schema_history</code>.
-     */
-    public final FlywaySchemaHistory FLYWAY_SCHEMA_HISTORY = FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY;
-
-    /**
-     * The table <code>minance.minance_category</code>.
+     * The table <code>minance_category</code>.
      */
     public final MinanceCategory MINANCE_CATEGORY = MinanceCategory.MINANCE_CATEGORY;
 
     /**
-     * The table <code>minance.raw_category_to_minance_category</code>.
+     * The table <code>raw_category_to_minance_category</code>.
      */
     public final RawCategoryToMinanceCategory RAW_CATEGORY_TO_MINANCE_CATEGORY = RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY;
 
     /**
-     * The table <code>minance.transactions</code>.
+     * The table <code>transactions</code>.
      */
     public final Transactions TRANSACTIONS = Transactions.TRANSACTIONS;
 
     /**
      * No further instances allowed
      */
-    private Minance() {
-        super("minance", null);
+    private DefaultSchema() {
+        super("", null);
     }
 
 
@@ -80,7 +74,6 @@ public class Minance extends SchemaImpl {
         return Arrays.asList(
             Accounts.ACCOUNTS,
             Banks.BANKS,
-            FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY,
             MinanceCategory.MINANCE_CATEGORY,
             RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY,
             Transactions.TRANSACTIONS
