@@ -6,45 +6,39 @@ package today.ihelio.jooq;
 
 import today.ihelio.jooq.tables.Accounts;
 import today.ihelio.jooq.tables.Banks;
-import today.ihelio.jooq.tables.FlywaySchemaHistory;
 import today.ihelio.jooq.tables.MinanceCategory;
 import today.ihelio.jooq.tables.RawCategoryToMinanceCategory;
 import today.ihelio.jooq.tables.Transactions;
 
 
 /**
- * Convenience access to all tables in minance.
+ * Convenience access to all tables in the default schema.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class Tables {
 
     /**
-     * The table <code>minance.accounts</code>.
+     * The table <code>accounts</code>.
      */
     public static final Accounts ACCOUNTS = Accounts.ACCOUNTS;
 
     /**
-     * The table <code>minance.banks</code>.
+     * The table <code>banks</code>.
      */
     public static final Banks BANKS = Banks.BANKS;
 
     /**
-     * The table <code>minance.flyway_schema_history</code>.
-     */
-    public static final FlywaySchemaHistory FLYWAY_SCHEMA_HISTORY = FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY;
-
-    /**
-     * The table <code>minance.minance_category</code>.
+     * The table <code>minance_category</code>.
      */
     public static final MinanceCategory MINANCE_CATEGORY = MinanceCategory.MINANCE_CATEGORY;
 
     /**
-     * The table <code>minance.raw_category_to_minance_category</code>.
+     * The table <code>raw_category_to_minance_category</code>.
      */
     public static final RawCategoryToMinanceCategory RAW_CATEGORY_TO_MINANCE_CATEGORY = RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY;
 
     /**
-     * The table <code>minance.transactions</code>.
+     * The table <code>transactions</code>.
      */
     public static final Transactions TRANSACTIONS = Transactions.TRANSACTIONS;
 }
