@@ -6,20 +6,23 @@ package today.ihelio.jooq.tables;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
-import java.util.function.Function;
 
+import org.jooq.Condition;
 import org.jooq.Field;
 import org.jooq.ForeignKey;
-import org.jooq.Function6;
 import org.jooq.Identity;
-import org.jooq.Index;
+import org.jooq.InverseForeignKey;
 import org.jooq.Name;
+import org.jooq.Path;
+import org.jooq.PlainSQL;
+import org.jooq.QueryPart;
 import org.jooq.Record;
-import org.jooq.Records;
-import org.jooq.Row6;
+import org.jooq.SQL;
 import org.jooq.Schema;
-import org.jooq.SelectField;
+import org.jooq.Select;
+import org.jooq.Stringly;
 import org.jooq.Table;
 import org.jooq.TableField;
 import org.jooq.TableOptions;
@@ -28,22 +31,23 @@ import org.jooq.impl.DSL;
 import org.jooq.impl.SQLDataType;
 import org.jooq.impl.TableImpl;
 
-import today.ihelio.jooq.Indexes;
+import today.ihelio.jooq.DefaultSchema;
 import today.ihelio.jooq.Keys;
-import today.ihelio.jooq.Minance;
+import today.ihelio.jooq.tables.Banks.BanksPath;
+import today.ihelio.jooq.tables.Transactions.TransactionsPath;
 import today.ihelio.jooq.tables.records.AccountsRecord;
 
 
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class Accounts extends TableImpl<AccountsRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * The reference instance of <code>minance.accounts</code>
+     * The reference instance of <code>accounts</code>
      */
     public static final Accounts ACCOUNTS = new Accounts();
 
@@ -56,76 +60,100 @@ public class Accounts extends TableImpl<AccountsRecord> {
     }
 
     /**
-     * The column <code>minance.accounts.account_id</code>.
+     * The column <code>accounts.account_id</code>.
      */
-    public final TableField<AccountsRecord, Integer> ACCOUNT_ID = createField(DSL.name("account_id"), SQLDataType.INTEGER.nullable(false).identity(true), this, "");
+    public final TableField<AccountsRecord, Integer> ACCOUNT_ID = createField(DSL.name("account_id"), SQLDataType.INTEGER.identity(true), this, "");
 
     /**
-     * The column <code>minance.accounts.bank_id</code>.
+     * The column <code>accounts.bank_id</code>.
      */
     public final TableField<AccountsRecord, Integer> BANK_ID = createField(DSL.name("bank_id"), SQLDataType.INTEGER.nullable(false), this, "");
 
     /**
-     * The column <code>minance.accounts.bank_name</code>.
+     * The column <code>accounts.bank_name</code>.
      */
-    public final TableField<AccountsRecord, String> BANK_NAME = createField(DSL.name("bank_name"), SQLDataType.VARCHAR(25).nullable(false), this, "");
+    public final TableField<AccountsRecord, String> BANK_NAME = createField(DSL.name("bank_name"), SQLDataType.CLOB.nullable(false), this, "");
 
     /**
-     * The column <code>minance.accounts.account_name</code>.
+     * The column <code>accounts.account_name</code>.
      */
-    public final TableField<AccountsRecord, String> ACCOUNT_NAME = createField(DSL.name("account_name"), SQLDataType.VARCHAR(25).nullable(false), this, "");
+    public final TableField<AccountsRecord, String> ACCOUNT_NAME = createField(DSL.name("account_name"), SQLDataType.CLOB.nullable(false), this, "");
 
     /**
-     * The column <code>minance.accounts.account_type</code>.
+     * The column <code>accounts.account_type</code>.
      */
-    public final TableField<AccountsRecord, String> ACCOUNT_TYPE = createField(DSL.name("account_type"), SQLDataType.VARCHAR(25).nullable(false), this, "");
+    public final TableField<AccountsRecord, String> ACCOUNT_TYPE = createField(DSL.name("account_type"), SQLDataType.CLOB.nullable(false), this, "");
 
     /**
-     * The column <code>minance.accounts.init_balance</code>.
+     * The column <code>accounts.init_balance</code>.
      */
-    public final TableField<AccountsRecord, BigDecimal> INIT_BALANCE = createField(DSL.name("init_balance"), SQLDataType.DECIMAL(9, 2), this, "");
+    public final TableField<AccountsRecord, BigDecimal> INIT_BALANCE = createField(DSL.name("init_balance"), SQLDataType.NUMERIC(9, 2), this, "");
 
     private Accounts(Name alias, Table<AccountsRecord> aliased) {
-        this(alias, aliased, null);
+        this(alias, aliased, (Field<?>[]) null, null);
     }
 
-    private Accounts(Name alias, Table<AccountsRecord> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table());
+    private Accounts(Name alias, Table<AccountsRecord> aliased, Field<?>[] parameters, Condition where) {
+        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
     }
 
     /**
-     * Create an aliased <code>minance.accounts</code> table reference
+     * Create an aliased <code>accounts</code> table reference
      */
     public Accounts(String alias) {
         this(DSL.name(alias), ACCOUNTS);
     }
 
     /**
-     * Create an aliased <code>minance.accounts</code> table reference
+     * Create an aliased <code>accounts</code> table reference
      */
     public Accounts(Name alias) {
         this(alias, ACCOUNTS);
     }
 
     /**
-     * Create a <code>minance.accounts</code> table reference
+     * Create a <code>accounts</code> table reference
      */
     public Accounts() {
         this(DSL.name("accounts"), null);
     }
 
-    public <O extends Record> Accounts(Table<O> child, ForeignKey<O, AccountsRecord> key) {
-        super(child, key, ACCOUNTS);
+    public <O extends Record> Accounts(Table<O> path, ForeignKey<O, AccountsRecord> childPath, InverseForeignKey<O, AccountsRecord> parentPath) {
+        super(path, childPath, parentPath, ACCOUNTS);
+    }
+
+    /**
+     * A subtype implementing {@link Path} for simplified path-based joins.
+     */
+    public static class AccountsPath extends Accounts implements Path<AccountsRecord> {
+
+        private static final long serialVersionUID = 1L;
+        public <O extends Record> AccountsPath(Table<O> path, ForeignKey<O, AccountsRecord> childPath, InverseForeignKey<O, AccountsRecord> parentPath) {
+            super(path, childPath, parentPath);
+        }
+        private AccountsPath(Name alias, Table<AccountsRecord> aliased) {
+            super(alias, aliased);
+        }
+
+        @Override
+        public AccountsPath as(String alias) {
+            return new AccountsPath(DSL.name(alias), this);
+        }
+
+        @Override
+        public AccountsPath as(Name alias) {
+            return new AccountsPath(alias, this);
+        }
+
+        @Override
+        public AccountsPath as(Table<?> alias) {
+            return new AccountsPath(alias.getQualifiedName(), this);
+        }
     }
 
     @Override
     public Schema getSchema() {
-        return aliased() ? null : Minance.MINANCE;
-    }
-
-    @Override
-    public List<Index> getIndexes() {
-        return Arrays.asList(Indexes.ACCOUNTS_BANK_ID);
+        return aliased() ? null : DefaultSchema.DEFAULT_SCHEMA;
     }
 
     @Override
@@ -135,29 +163,41 @@ public class Accounts extends TableImpl<AccountsRecord> {
 
     @Override
     public UniqueKey<AccountsRecord> getPrimaryKey() {
-        return Keys.KEY_ACCOUNTS_PRIMARY;
+        return Keys.ACCOUNTS__PK_ACCOUNTS;
     }
 
     @Override
     public List<UniqueKey<AccountsRecord>> getUniqueKeys() {
-        return Arrays.asList(Keys.KEY_ACCOUNTS_ACCOUNT_UNIQUE);
+        return Arrays.asList(Keys.ACCOUNTS__UK_ACCOUNTS_70892193);
     }
 
     @Override
     public List<ForeignKey<AccountsRecord, ?>> getReferences() {
-        return Arrays.asList(Keys.ACCOUNTS_IBFK_1);
+        return Arrays.asList(Keys.ACCOUNTS__FK_ACCOUNTS_PK_BANKS);
     }
 
-    private transient Banks _banks;
+    private transient BanksPath _banks;
 
     /**
-     * Get the implicit join path to the <code>minance.banks</code> table.
+     * Get the implicit join path to the <code>banks</code> table.
      */
-    public Banks banks() {
+    public BanksPath banks() {
         if (_banks == null)
-            _banks = new Banks(this, Keys.ACCOUNTS_IBFK_1);
+            _banks = new BanksPath(this, Keys.ACCOUNTS__FK_ACCOUNTS_PK_BANKS, null);
 
         return _banks;
+    }
+
+    private transient TransactionsPath _transactions;
+
+    /**
+     * Get the implicit to-many join path to the <code>transactions</code> table
+     */
+    public TransactionsPath transactions() {
+        if (_transactions == null)
+            _transactions = new TransactionsPath(this, null, Keys.TRANSACTIONS__FK_TRANSACTIONS_PK_ACCOUNTS.getInverseKey());
+
+        return _transactions;
     }
 
     @Override
@@ -199,27 +239,87 @@ public class Accounts extends TableImpl<AccountsRecord> {
         return new Accounts(name.getQualifiedName(), null);
     }
 
-    // -------------------------------------------------------------------------
-    // Row6 type methods
-    // -------------------------------------------------------------------------
-
+    /**
+     * Create an inline derived table from this table
+     */
     @Override
-    public Row6<Integer, Integer, String, String, String, BigDecimal> fieldsRow() {
-        return (Row6) super.fieldsRow();
+    public Accounts where(Condition condition) {
+        return new Accounts(getQualifiedName(), aliased() ? this : null, null, condition);
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Function6<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super BigDecimal, ? extends U> from) {
-        return convertFrom(Records.mapping(from));
+    @Override
+    public Accounts where(Collection<? extends Condition> conditions) {
+        return where(DSL.and(conditions));
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Class,
-     * Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Class<U> toType, Function6<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super BigDecimal, ? extends U> from) {
-        return convertFrom(toType, Records.mapping(from));
+    @Override
+    public Accounts where(Condition... conditions) {
+        return where(DSL.and(conditions));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Accounts where(Field<Boolean> condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Accounts where(SQL condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Accounts where(@Stringly.SQL String condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Accounts where(@Stringly.SQL String condition, Object... binds) {
+        return where(DSL.condition(condition, binds));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Accounts where(@Stringly.SQL String condition, QueryPart... parts) {
+        return where(DSL.condition(condition, parts));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Accounts whereExists(Select<?> select) {
+        return where(DSL.exists(select));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Accounts whereNotExists(Select<?> select) {
+        return where(DSL.notExists(select));
     }
 }
