@@ -5,20 +5,23 @@ package today.ihelio.jooq.tables;
 
 
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
-import java.util.function.Function;
 
+import org.jooq.Condition;
 import org.jooq.Field;
 import org.jooq.ForeignKey;
-import org.jooq.Function3;
 import org.jooq.Identity;
-import org.jooq.Index;
+import org.jooq.InverseForeignKey;
 import org.jooq.Name;
+import org.jooq.Path;
+import org.jooq.PlainSQL;
+import org.jooq.QueryPart;
 import org.jooq.Record;
-import org.jooq.Records;
-import org.jooq.Row3;
+import org.jooq.SQL;
 import org.jooq.Schema;
-import org.jooq.SelectField;
+import org.jooq.Select;
+import org.jooq.Stringly;
 import org.jooq.Table;
 import org.jooq.TableField;
 import org.jooq.TableOptions;
@@ -27,23 +30,22 @@ import org.jooq.impl.DSL;
 import org.jooq.impl.SQLDataType;
 import org.jooq.impl.TableImpl;
 
-import today.ihelio.jooq.Indexes;
+import today.ihelio.jooq.DefaultSchema;
 import today.ihelio.jooq.Keys;
-import today.ihelio.jooq.Minance;
+import today.ihelio.jooq.tables.MinanceCategory.MinanceCategoryPath;
 import today.ihelio.jooq.tables.records.RawCategoryToMinanceCategoryRecord;
 
 
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class RawCategoryToMinanceCategory extends TableImpl<RawCategoryToMinanceCategoryRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * The reference instance of
-     * <code>minance.raw_category_to_minance_category</code>
+     * The reference instance of <code>raw_category_to_minance_category</code>
      */
     public static final RawCategoryToMinanceCategory RAW_CATEGORY_TO_MINANCE_CATEGORY = new RawCategoryToMinanceCategory();
 
@@ -56,67 +58,88 @@ public class RawCategoryToMinanceCategory extends TableImpl<RawCategoryToMinance
     }
 
     /**
-     * The column
-     * <code>minance.raw_category_to_minance_category.rc_to_mc_id</code>.
+     * The column <code>raw_category_to_minance_category.rc_to_mc_id</code>.
      */
-    public final TableField<RawCategoryToMinanceCategoryRecord, Integer> RC_TO_MC_ID = createField(DSL.name("rc_to_mc_id"), SQLDataType.INTEGER.nullable(false).identity(true), this, "");
+    public final TableField<RawCategoryToMinanceCategoryRecord, Integer> RC_TO_MC_ID = createField(DSL.name("rc_to_mc_id"), SQLDataType.INTEGER.identity(true), this, "");
+
+    /**
+     * The column <code>raw_category_to_minance_category.raw_category</code>.
+     */
+    public final TableField<RawCategoryToMinanceCategoryRecord, String> RAW_CATEGORY = createField(DSL.name("raw_category"), SQLDataType.CLOB.nullable(false), this, "");
 
     /**
      * The column
-     * <code>minance.raw_category_to_minance_category.raw_category</code>.
-     */
-    public final TableField<RawCategoryToMinanceCategoryRecord, String> RAW_CATEGORY = createField(DSL.name("raw_category"), SQLDataType.VARCHAR(100).nullable(false), this, "");
-
-    /**
-     * The column
-     * <code>minance.raw_category_to_minance_category.minance_category_id</code>.
+     * <code>raw_category_to_minance_category.minance_category_id</code>.
      */
     public final TableField<RawCategoryToMinanceCategoryRecord, Integer> MINANCE_CATEGORY_ID = createField(DSL.name("minance_category_id"), SQLDataType.INTEGER.nullable(false), this, "");
 
     private RawCategoryToMinanceCategory(Name alias, Table<RawCategoryToMinanceCategoryRecord> aliased) {
-        this(alias, aliased, null);
+        this(alias, aliased, (Field<?>[]) null, null);
     }
 
-    private RawCategoryToMinanceCategory(Name alias, Table<RawCategoryToMinanceCategoryRecord> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table());
+    private RawCategoryToMinanceCategory(Name alias, Table<RawCategoryToMinanceCategoryRecord> aliased, Field<?>[] parameters, Condition where) {
+        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
     }
 
     /**
-     * Create an aliased <code>minance.raw_category_to_minance_category</code>
-     * table reference
+     * Create an aliased <code>raw_category_to_minance_category</code> table
+     * reference
      */
     public RawCategoryToMinanceCategory(String alias) {
         this(DSL.name(alias), RAW_CATEGORY_TO_MINANCE_CATEGORY);
     }
 
     /**
-     * Create an aliased <code>minance.raw_category_to_minance_category</code>
-     * table reference
+     * Create an aliased <code>raw_category_to_minance_category</code> table
+     * reference
      */
     public RawCategoryToMinanceCategory(Name alias) {
         this(alias, RAW_CATEGORY_TO_MINANCE_CATEGORY);
     }
 
     /**
-     * Create a <code>minance.raw_category_to_minance_category</code> table
-     * reference
+     * Create a <code>raw_category_to_minance_category</code> table reference
      */
     public RawCategoryToMinanceCategory() {
         this(DSL.name("raw_category_to_minance_category"), null);
     }
 
-    public <O extends Record> RawCategoryToMinanceCategory(Table<O> child, ForeignKey<O, RawCategoryToMinanceCategoryRecord> key) {
-        super(child, key, RAW_CATEGORY_TO_MINANCE_CATEGORY);
+    public <O extends Record> RawCategoryToMinanceCategory(Table<O> path, ForeignKey<O, RawCategoryToMinanceCategoryRecord> childPath, InverseForeignKey<O, RawCategoryToMinanceCategoryRecord> parentPath) {
+        super(path, childPath, parentPath, RAW_CATEGORY_TO_MINANCE_CATEGORY);
+    }
+
+    /**
+     * A subtype implementing {@link Path} for simplified path-based joins.
+     */
+    public static class RawCategoryToMinanceCategoryPath extends RawCategoryToMinanceCategory implements Path<RawCategoryToMinanceCategoryRecord> {
+
+        private static final long serialVersionUID = 1L;
+        public <O extends Record> RawCategoryToMinanceCategoryPath(Table<O> path, ForeignKey<O, RawCategoryToMinanceCategoryRecord> childPath, InverseForeignKey<O, RawCategoryToMinanceCategoryRecord> parentPath) {
+            super(path, childPath, parentPath);
+        }
+        private RawCategoryToMinanceCategoryPath(Name alias, Table<RawCategoryToMinanceCategoryRecord> aliased) {
+            super(alias, aliased);
+        }
+
+        @Override
+        public RawCategoryToMinanceCategoryPath as(String alias) {
+            return new RawCategoryToMinanceCategoryPath(DSL.name(alias), this);
+        }
+
+        @Override
+        public RawCategoryToMinanceCategoryPath as(Name alias) {
+            return new RawCategoryToMinanceCategoryPath(alias, this);
+        }
+
+        @Override
+        public RawCategoryToMinanceCategoryPath as(Table<?> alias) {
+            return new RawCategoryToMinanceCategoryPath(alias.getQualifiedName(), this);
+        }
     }
 
     @Override
     public Schema getSchema() {
-        return aliased() ? null : Minance.MINANCE;
-    }
-
-    @Override
-    public List<Index> getIndexes() {
-        return Arrays.asList(Indexes.RAW_CATEGORY_TO_MINANCE_CATEGORY_MINANCE_CATEGORY_ID);
+        return aliased() ? null : DefaultSchema.DEFAULT_SCHEMA;
     }
 
     @Override
@@ -126,23 +149,22 @@ public class RawCategoryToMinanceCategory extends TableImpl<RawCategoryToMinance
 
     @Override
     public UniqueKey<RawCategoryToMinanceCategoryRecord> getPrimaryKey() {
-        return Keys.KEY_RAW_CATEGORY_TO_MINANCE_CATEGORY_PRIMARY;
+        return Keys.RAW_CATEGORY_TO_MINANCE_CATEGORY__PK_RAW_CATEGORY_TO_MINANCE_CATEGORY;
     }
 
     @Override
     public List<ForeignKey<RawCategoryToMinanceCategoryRecord, ?>> getReferences() {
-        return Arrays.asList(Keys.RAW_CATEGORY_TO_MINANCE_CATEGORY_IBFK_1);
+        return Arrays.asList(Keys.RAW_CATEGORY_TO_MINANCE_CATEGORY__FK_RAW_CATEGORY_TO_MINANCE_CATEGORY_PK_MINANCE_CATEGORY);
     }
 
-    private transient MinanceCategory _minanceCategory;
+    private transient MinanceCategoryPath _minanceCategory;
 
     /**
-     * Get the implicit join path to the <code>minance.minance_category</code>
-     * table.
+     * Get the implicit join path to the <code>minance_category</code> table.
      */
-    public MinanceCategory minanceCategory() {
+    public MinanceCategoryPath minanceCategory() {
         if (_minanceCategory == null)
-            _minanceCategory = new MinanceCategory(this, Keys.RAW_CATEGORY_TO_MINANCE_CATEGORY_IBFK_1);
+            _minanceCategory = new MinanceCategoryPath(this, Keys.RAW_CATEGORY_TO_MINANCE_CATEGORY__FK_RAW_CATEGORY_TO_MINANCE_CATEGORY_PK_MINANCE_CATEGORY, null);
 
         return _minanceCategory;
     }
@@ -186,27 +208,87 @@ public class RawCategoryToMinanceCategory extends TableImpl<RawCategoryToMinance
         return new RawCategoryToMinanceCategory(name.getQualifiedName(), null);
     }
 
-    // -------------------------------------------------------------------------
-    // Row3 type methods
-    // -------------------------------------------------------------------------
-
+    /**
+     * Create an inline derived table from this table
+     */
     @Override
-    public Row3<Integer, String, Integer> fieldsRow() {
-        return (Row3) super.fieldsRow();
+    public RawCategoryToMinanceCategory where(Condition condition) {
+        return new RawCategoryToMinanceCategory(getQualifiedName(), aliased() ? this : null, null, condition);
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Function3<? super Integer, ? super String, ? super Integer, ? extends U> from) {
-        return convertFrom(Records.mapping(from));
+    @Override
+    public RawCategoryToMinanceCategory where(Collection<? extends Condition> conditions) {
+        return where(DSL.and(conditions));
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Class,
-     * Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Class<U> toType, Function3<? super Integer, ? super String, ? super Integer, ? extends U> from) {
-        return convertFrom(toType, Records.mapping(from));
+    @Override
+    public RawCategoryToMinanceCategory where(Condition... conditions) {
+        return where(DSL.and(conditions));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public RawCategoryToMinanceCategory where(Field<Boolean> condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public RawCategoryToMinanceCategory where(SQL condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public RawCategoryToMinanceCategory where(@Stringly.SQL String condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public RawCategoryToMinanceCategory where(@Stringly.SQL String condition, Object... binds) {
+        return where(DSL.condition(condition, binds));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public RawCategoryToMinanceCategory where(@Stringly.SQL String condition, QueryPart... parts) {
+        return where(DSL.condition(condition, parts));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public RawCategoryToMinanceCategory whereExists(Select<?> select) {
+        return where(DSL.exists(select));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public RawCategoryToMinanceCategory whereNotExists(Select<?> select) {
+        return where(DSL.notExists(select));
     }
 }
