@@ -7,19 +7,23 @@ package today.ihelio.jooq.tables;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
-import java.util.function.Function;
 
+import org.jooq.Condition;
 import org.jooq.Field;
 import org.jooq.ForeignKey;
-import org.jooq.Function18;
 import org.jooq.Identity;
+import org.jooq.InverseForeignKey;
 import org.jooq.Name;
+import org.jooq.Path;
+import org.jooq.PlainSQL;
+import org.jooq.QueryPart;
 import org.jooq.Record;
-import org.jooq.Records;
-import org.jooq.Row18;
+import org.jooq.SQL;
 import org.jooq.Schema;
-import org.jooq.SelectField;
+import org.jooq.Select;
+import org.jooq.Stringly;
 import org.jooq.Table;
 import org.jooq.TableField;
 import org.jooq.TableOptions;
@@ -28,21 +32,22 @@ import org.jooq.impl.DSL;
 import org.jooq.impl.SQLDataType;
 import org.jooq.impl.TableImpl;
 
+import today.ihelio.jooq.DefaultSchema;
 import today.ihelio.jooq.Keys;
-import today.ihelio.jooq.Minance;
+import today.ihelio.jooq.tables.Accounts.AccountsPath;
 import today.ihelio.jooq.tables.records.TransactionsRecord;
 
 
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class Transactions extends TableImpl<TransactionsRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * The reference instance of <code>minance.transactions</code>
+     * The reference instance of <code>transactions</code>
      */
     public static final Transactions TRANSACTIONS = new Transactions();
 
@@ -55,131 +60,160 @@ public class Transactions extends TableImpl<TransactionsRecord> {
     }
 
     /**
-     * The column <code>minance.transactions.transaction_id</code>.
+     * The column <code>transactions.transaction_id</code>.
      */
-    public final TableField<TransactionsRecord, Integer> TRANSACTION_ID = createField(DSL.name("transaction_id"), SQLDataType.INTEGER.nullable(false).identity(true), this, "");
+    public final TableField<TransactionsRecord, Integer> TRANSACTION_ID = createField(DSL.name("transaction_id"), SQLDataType.INTEGER.identity(true), this, "");
 
     /**
-     * The column <code>minance.transactions.account_id</code>.
+     * The column <code>transactions.account_id</code>.
      */
     public final TableField<TransactionsRecord, Integer> ACCOUNT_ID = createField(DSL.name("account_id"), SQLDataType.INTEGER, this, "");
 
     /**
-     * The column <code>minance.transactions.category</code>.
+     * The column <code>transactions.category</code>.
      */
-    public final TableField<TransactionsRecord, String> CATEGORY = createField(DSL.name("category"), SQLDataType.VARCHAR(50), this, "");
+    public final TableField<TransactionsRecord, String> CATEGORY = createField(DSL.name("category"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.description</code>.
+     * The column <code>transactions.description</code>.
      */
-    public final TableField<TransactionsRecord, String> DESCRIPTION = createField(DSL.name("description"), SQLDataType.VARCHAR(100), this, "");
+    public final TableField<TransactionsRecord, String> DESCRIPTION = createField(DSL.name("description"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.transaction_type</code>.
+     * The column <code>transactions.transaction_type</code>.
      */
-    public final TableField<TransactionsRecord, String> TRANSACTION_TYPE = createField(DSL.name("transaction_type"), SQLDataType.VARCHAR(50), this, "");
+    public final TableField<TransactionsRecord, String> TRANSACTION_TYPE = createField(DSL.name("transaction_type"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.transaction_date</code>.
+     * The column <code>transactions.transaction_date</code>.
      */
     public final TableField<TransactionsRecord, LocalDate> TRANSACTION_DATE = createField(DSL.name("transaction_date"), SQLDataType.LOCALDATE, this, "");
 
     /**
-     * The column <code>minance.transactions.post_date</code>.
+     * The column <code>transactions.post_date</code>.
      */
     public final TableField<TransactionsRecord, LocalDate> POST_DATE = createField(DSL.name("post_date"), SQLDataType.LOCALDATE, this, "");
 
     /**
-     * The column <code>minance.transactions.memo</code>.
+     * The column <code>transactions.memo</code>.
      */
-    public final TableField<TransactionsRecord, String> MEMO = createField(DSL.name("memo"), SQLDataType.VARCHAR(200), this, "");
+    public final TableField<TransactionsRecord, String> MEMO = createField(DSL.name("memo"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.address</code>.
+     * The column <code>transactions.address</code>.
      */
-    public final TableField<TransactionsRecord, String> ADDRESS = createField(DSL.name("address"), SQLDataType.VARCHAR(100), this, "");
+    public final TableField<TransactionsRecord, String> ADDRESS = createField(DSL.name("address"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.city</code>.
+     * The column <code>transactions.city</code>.
      */
-    public final TableField<TransactionsRecord, String> CITY = createField(DSL.name("city"), SQLDataType.VARCHAR(30), this, "");
+    public final TableField<TransactionsRecord, String> CITY = createField(DSL.name("city"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.state_name</code>.
+     * The column <code>transactions.state_name</code>.
      */
-    public final TableField<TransactionsRecord, String> STATE_NAME = createField(DSL.name("state_name"), SQLDataType.VARCHAR(30), this, "");
+    public final TableField<TransactionsRecord, String> STATE_NAME = createField(DSL.name("state_name"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.country</code>.
+     * The column <code>transactions.country</code>.
      */
-    public final TableField<TransactionsRecord, String> COUNTRY = createField(DSL.name("country"), SQLDataType.VARCHAR(10), this, "");
+    public final TableField<TransactionsRecord, String> COUNTRY = createField(DSL.name("country"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.zipcode</code>.
+     * The column <code>transactions.zipcode</code>.
      */
-    public final TableField<TransactionsRecord, String> ZIPCODE = createField(DSL.name("zipcode"), SQLDataType.VARCHAR(10), this, "");
+    public final TableField<TransactionsRecord, String> ZIPCODE = createField(DSL.name("zipcode"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.amount</code>.
+     * The column <code>transactions.amount</code>.
      */
-    public final TableField<TransactionsRecord, BigDecimal> AMOUNT = createField(DSL.name("amount"), SQLDataType.DECIMAL(9, 2), this, "");
+    public final TableField<TransactionsRecord, BigDecimal> AMOUNT = createField(DSL.name("amount"), SQLDataType.NUMERIC(9, 2), this, "");
 
     /**
-     * The column <code>minance.transactions.bank_name</code>.
+     * The column <code>transactions.bank_name</code>.
      */
-    public final TableField<TransactionsRecord, String> BANK_NAME = createField(DSL.name("bank_name"), SQLDataType.VARCHAR(50), this, "");
+    public final TableField<TransactionsRecord, String> BANK_NAME = createField(DSL.name("bank_name"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.account_name</code>.
+     * The column <code>transactions.account_name</code>.
      */
-    public final TableField<TransactionsRecord, String> ACCOUNT_NAME = createField(DSL.name("account_name"), SQLDataType.VARCHAR(50), this, "");
+    public final TableField<TransactionsRecord, String> ACCOUNT_NAME = createField(DSL.name("account_name"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.upload_time</code>.
+     * The column <code>transactions.upload_time</code>.
      */
-    public final TableField<TransactionsRecord, String> UPLOAD_TIME = createField(DSL.name("upload_time"), SQLDataType.VARCHAR(20), this, "");
+    public final TableField<TransactionsRecord, String> UPLOAD_TIME = createField(DSL.name("upload_time"), SQLDataType.CLOB, this, "");
 
     /**
-     * The column <code>minance.transactions.is_duplicate</code>.
+     * The column <code>transactions.is_duplicate</code>.
      */
-    public final TableField<TransactionsRecord, Boolean> IS_DUPLICATE = createField(DSL.name("is_duplicate"), SQLDataType.BOOLEAN.defaultValue(DSL.inline("0", SQLDataType.BOOLEAN)), this, "");
+    public final TableField<TransactionsRecord, Integer> IS_DUPLICATE = createField(DSL.name("is_duplicate"), SQLDataType.INTEGER.defaultValue(DSL.field(DSL.raw("0"), SQLDataType.INTEGER)), this, "");
 
     private Transactions(Name alias, Table<TransactionsRecord> aliased) {
-        this(alias, aliased, null);
+        this(alias, aliased, (Field<?>[]) null, null);
     }
 
-    private Transactions(Name alias, Table<TransactionsRecord> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table());
+    private Transactions(Name alias, Table<TransactionsRecord> aliased, Field<?>[] parameters, Condition where) {
+        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
     }
 
     /**
-     * Create an aliased <code>minance.transactions</code> table reference
+     * Create an aliased <code>transactions</code> table reference
      */
     public Transactions(String alias) {
         this(DSL.name(alias), TRANSACTIONS);
     }
 
     /**
-     * Create an aliased <code>minance.transactions</code> table reference
+     * Create an aliased <code>transactions</code> table reference
      */
     public Transactions(Name alias) {
         this(alias, TRANSACTIONS);
     }
 
     /**
-     * Create a <code>minance.transactions</code> table reference
+     * Create a <code>transactions</code> table reference
      */
     public Transactions() {
         this(DSL.name("transactions"), null);
     }
 
-    public <O extends Record> Transactions(Table<O> child, ForeignKey<O, TransactionsRecord> key) {
-        super(child, key, TRANSACTIONS);
+    public <O extends Record> Transactions(Table<O> path, ForeignKey<O, TransactionsRecord> childPath, InverseForeignKey<O, TransactionsRecord> parentPath) {
+        super(path, childPath, parentPath, TRANSACTIONS);
+    }
+
+    /**
+     * A subtype implementing {@link Path} for simplified path-based joins.
+     */
+    public static class TransactionsPath extends Transactions implements Path<TransactionsRecord> {
+
+        private static final long serialVersionUID = 1L;
+        public <O extends Record> TransactionsPath(Table<O> path, ForeignKey<O, TransactionsRecord> childPath, InverseForeignKey<O, TransactionsRecord> parentPath) {
+            super(path, childPath, parentPath);
+        }
+        private TransactionsPath(Name alias, Table<TransactionsRecord> aliased) {
+            super(alias, aliased);
+        }
+
+        @Override
+        public TransactionsPath as(String alias) {
+            return new TransactionsPath(DSL.name(alias), this);
+        }
+
+        @Override
+        public TransactionsPath as(Name alias) {
+            return new TransactionsPath(alias, this);
+        }
+
+        @Override
+        public TransactionsPath as(Table<?> alias) {
+            return new TransactionsPath(alias.getQualifiedName(), this);
+        }
     }
 
     @Override
     public Schema getSchema() {
-        return aliased() ? null : Minance.MINANCE;
+        return aliased() ? null : DefaultSchema.DEFAULT_SCHEMA;
     }
 
     @Override
@@ -189,27 +223,27 @@ public class Transactions extends TableImpl<TransactionsRecord> {
 
     @Override
     public UniqueKey<TransactionsRecord> getPrimaryKey() {
-        return Keys.KEY_TRANSACTIONS_PRIMARY;
+        return Keys.TRANSACTIONS__PK_TRANSACTIONS;
     }
 
     @Override
     public List<UniqueKey<TransactionsRecord>> getUniqueKeys() {
-        return Arrays.asList(Keys.KEY_TRANSACTIONS_TRANSACTION_UNIQUE);
+        return Arrays.asList(Keys.TRANSACTIONS__UK_TRANSACTIONS_6933421);
     }
 
     @Override
     public List<ForeignKey<TransactionsRecord, ?>> getReferences() {
-        return Arrays.asList(Keys.TRANSACTIONS_IBFK_1);
+        return Arrays.asList(Keys.TRANSACTIONS__FK_TRANSACTIONS_PK_ACCOUNTS);
     }
 
-    private transient Accounts _accounts;
+    private transient AccountsPath _accounts;
 
     /**
-     * Get the implicit join path to the <code>minance.accounts</code> table.
+     * Get the implicit join path to the <code>accounts</code> table.
      */
-    public Accounts accounts() {
+    public AccountsPath accounts() {
         if (_accounts == null)
-            _accounts = new Accounts(this, Keys.TRANSACTIONS_IBFK_1);
+            _accounts = new AccountsPath(this, Keys.TRANSACTIONS__FK_TRANSACTIONS_PK_ACCOUNTS, null);
 
         return _accounts;
     }
@@ -253,27 +287,87 @@ public class Transactions extends TableImpl<TransactionsRecord> {
         return new Transactions(name.getQualifiedName(), null);
     }
 
-    // -------------------------------------------------------------------------
-    // Row18 type methods
-    // -------------------------------------------------------------------------
-
+    /**
+     * Create an inline derived table from this table
+     */
     @Override
-    public Row18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Boolean> fieldsRow() {
-        return (Row18) super.fieldsRow();
+    public Transactions where(Condition condition) {
+        return new Transactions(getQualifiedName(), aliased() ? this : null, null, condition);
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Function18<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super LocalDate, ? super LocalDate, ? super String, ? super String, ? super String, ? super String, ? super String, ? super String, ? super BigDecimal, ? super String, ? super String, ? super String, ? super Boolean, ? extends U> from) {
-        return convertFrom(Records.mapping(from));
+    @Override
+    public Transactions where(Collection<? extends Condition> conditions) {
+        return where(DSL.and(conditions));
     }
 
     /**
-     * Convenience mapping calling {@link SelectField#convertFrom(Class,
-     * Function)}.
+     * Create an inline derived table from this table
      */
-    public <U> SelectField<U> mapping(Class<U> toType, Function18<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super LocalDate, ? super LocalDate, ? super String, ? super String, ? super String, ? super String, ? super String, ? super String, ? super BigDecimal, ? super String, ? super String, ? super String, ? super Boolean, ? extends U> from) {
-        return convertFrom(toType, Records.mapping(from));
+    @Override
+    public Transactions where(Condition... conditions) {
+        return where(DSL.and(conditions));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions where(Field<Boolean> condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(SQL condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition, Object... binds) {
+        return where(DSL.condition(condition, binds));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition, QueryPart... parts) {
+        return where(DSL.condition(condition, parts));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions whereExists(Select<?> select) {
+        return where(DSL.exists(select));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions whereNotExists(Select<?> select) {
+        return where(DSL.notExists(select));
     }
 }
