@@ -18,7 +18,7 @@ import today.ihelio.jooq.tables.records.AccountsRecord;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class AccountsDao extends DAOImpl<AccountsRecord, today.ihelio.jooq.tables.pojos.Accounts, Integer> {
 
     /**
