@@ -17,7 +17,7 @@ import today.ihelio.jooq.tables.records.BanksRecord;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class BanksDao extends DAOImpl<BanksRecord, today.ihelio.jooq.tables.pojos.Banks, Integer> {
 
     /**
