@@ -17,7 +17,7 @@ import today.ihelio.jooq.tables.records.MinanceCategoryRecord;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class MinanceCategoryDao extends DAOImpl<MinanceCategoryRecord, today.ihelio.jooq.tables.pojos.MinanceCategory, Integer> {
 
     /**
