@@ -17,7 +17,7 @@ import today.ihelio.jooq.tables.records.RawCategoryToMinanceCategoryRecord;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class RawCategoryToMinanceCategoryDao extends DAOImpl<RawCategoryToMinanceCategoryRecord, today.ihelio.jooq.tables.pojos.RawCategoryToMinanceCategory, Integer> {
 
     /**
