@@ -12,7 +12,7 @@ import java.time.LocalDate;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
 public class Transactions implements Serializable {
 
     private static final long serialVersionUID = 1L;
@@ -34,7 +34,7 @@ public class Transactions implements Serializable {
     private String bankName;
     private String accountName;
     private String uploadTime;
-    private Boolean isDuplicate;
+    private Integer isDuplicate;
 
     public Transactions() {}
 
@@ -77,7 +77,7 @@ public class Transactions implements Serializable {
         String bankName,
         String accountName,
         String uploadTime,
-        Boolean isDuplicate
+        Integer isDuplicate
     ) {
         this.transactionId = transactionId;
         this.accountId = accountId;
@@ -100,254 +100,254 @@ public class Transactions implements Serializable {
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_id</code>.
+     * Getter for <code>transactions.transaction_id</code>.
      */
     public Integer getTransactionId() {
         return this.transactionId;
     }
 
     /**
-     * Setter for <code>minance.transactions.transaction_id</code>.
+     * Setter for <code>transactions.transaction_id</code>.
      */
     public void setTransactionId(Integer transactionId) {
         this.transactionId = transactionId;
     }
 
     /**
-     * Getter for <code>minance.transactions.account_id</code>.
+     * Getter for <code>transactions.account_id</code>.
      */
     public Integer getAccountId() {
         return this.accountId;
     }
 
     /**
-     * Setter for <code>minance.transactions.account_id</code>.
+     * Setter for <code>transactions.account_id</code>.
      */
     public void setAccountId(Integer accountId) {
         this.accountId = accountId;
     }
 
     /**
-     * Getter for <code>minance.transactions.category</code>.
+     * Getter for <code>transactions.category</code>.
      */
     public String getCategory() {
         return this.category;
     }
 
     /**
-     * Setter for <code>minance.transactions.category</code>.
+     * Setter for <code>transactions.category</code>.
      */
     public void setCategory(String category) {
         this.category = category;
     }
 
     /**
-     * Getter for <code>minance.transactions.description</code>.
+     * Getter for <code>transactions.description</code>.
      */
     public String getDescription() {
         return this.description;
     }
 
     /**
-     * Setter for <code>minance.transactions.description</code>.
+     * Setter for <code>transactions.description</code>.
      */
     public void setDescription(String description) {
         this.description = description;
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_type</code>.
+     * Getter for <code>transactions.transaction_type</code>.
      */
     public String getTransactionType() {
         return this.transactionType;
     }
 
     /**
-     * Setter for <code>minance.transactions.transaction_type</code>.
+     * Setter for <code>transactions.transaction_type</code>.
      */
     public void setTransactionType(String transactionType) {
         this.transactionType = transactionType;
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_date</code>.
+     * Getter for <code>transactions.transaction_date</code>.
      */
     public LocalDate getTransactionDate() {
         return this.transactionDate;
     }
 
     /**
-     * Setter for <code>minance.transactions.transaction_date</code>.
+     * Setter for <code>transactions.transaction_date</code>.
      */
     public void setTransactionDate(LocalDate transactionDate) {
         this.transactionDate = transactionDate;
     }
 
     /**
-     * Getter for <code>minance.transactions.post_date</code>.
+     * Getter for <code>transactions.post_date</code>.
      */
     public LocalDate getPostDate() {
         return this.postDate;
     }
 
     /**
-     * Setter for <code>minance.transactions.post_date</code>.
+     * Setter for <code>transactions.post_date</code>.
      */
     public void setPostDate(LocalDate postDate) {
         this.postDate = postDate;
     }
 
     /**
-     * Getter for <code>minance.transactions.memo</code>.
+     * Getter for <code>transactions.memo</code>.
      */
     public String getMemo() {
         return this.memo;
     }
 
     /**
-     * Setter for <code>minance.transactions.memo</code>.
+     * Setter for <code>transactions.memo</code>.
      */
     public void setMemo(String memo) {
         this.memo = memo;
     }
 
     /**
-     * Getter for <code>minance.transactions.address</code>.
+     * Getter for <code>transactions.address</code>.
      */
     public String getAddress() {
         return this.address;
     }
 
     /**
-     * Setter for <code>minance.transactions.address</code>.
+     * Setter for <code>transactions.address</code>.
      */
     public void setAddress(String address) {
         this.address = address;
     }
 
     /**
-     * Getter for <code>minance.transactions.city</code>.
+     * Getter for <code>transactions.city</code>.
      */
     public String getCity() {
         return this.city;
     }
 
     /**
-     * Setter for <code>minance.transactions.city</code>.
+     * Setter for <code>transactions.city</code>.
      */
     public void setCity(String city) {
         this.city = city;
     }
 
     /**
-     * Getter for <code>minance.transactions.state_name</code>.
+     * Getter for <code>transactions.state_name</code>.
      */
     public String getStateName() {
         return this.stateName;
     }
 
     /**
-     * Setter for <code>minance.transactions.state_name</code>.
+     * Setter for <code>transactions.state_name</code>.
      */
     public void setStateName(String stateName) {
         this.stateName = stateName;
     }
 
     /**
-     * Getter for <code>minance.transactions.country</code>.
+     * Getter for <code>transactions.country</code>.
      */
     public String getCountry() {
         return this.country;
     }
 
     /**
-     * Setter for <code>minance.transactions.country</code>.
+     * Setter for <code>transactions.country</code>.
      */
     public void setCountry(String country) {
         this.country = country;
     }
 
     /**
-     * Getter for <code>minance.transactions.zipcode</code>.
+     * Getter for <code>transactions.zipcode</code>.
      */
     public String getZipcode() {
         return this.zipcode;
     }
 
     /**
-     * Setter for <code>minance.transactions.zipcode</code>.
+     * Setter for <code>transactions.zipcode</code>.
      */
     public void setZipcode(String zipcode) {
         this.zipcode = zipcode;
     }
 
     /**
-     * Getter for <code>minance.transactions.amount</code>.
+     * Getter for <code>transactions.amount</code>.
      */
     public BigDecimal getAmount() {
         return this.amount;
     }
 
     /**
-     * Setter for <code>minance.transactions.amount</code>.
+     * Setter for <code>transactions.amount</code>.
      */
     public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
     /**
-     * Getter for <code>minance.transactions.bank_name</code>.
+     * Getter for <code>transactions.bank_name</code>.
      */
     public String getBankName() {
         return this.bankName;
     }
 
     /**
-     * Setter for <code>minance.transactions.bank_name</code>.
+     * Setter for <code>transactions.bank_name</code>.
      */
     public void setBankName(String bankName) {
         this.bankName = bankName;
     }
 
     /**
-     * Getter for <code>minance.transactions.account_name</code>.
+     * Getter for <code>transactions.account_name</code>.
      */
     public String getAccountName() {
         return this.accountName;
     }
 
     /**
-     * Setter for <code>minance.transactions.account_name</code>.
+     * Setter for <code>transactions.account_name</code>.
      */
     public void setAccountName(String accountName) {
         this.accountName = accountName;
     }
 
     /**
-     * Getter for <code>minance.transactions.upload_time</code>.
+     * Getter for <code>transactions.upload_time</code>.
      */
     public String getUploadTime() {
         return this.uploadTime;
     }
 
     /**
-     * Setter for <code>minance.transactions.upload_time</code>.
+     * Setter for <code>transactions.upload_time</code>.
      */
     public void setUploadTime(String uploadTime) {
         this.uploadTime = uploadTime;
     }
 
     /**
-     * Getter for <code>minance.transactions.is_duplicate</code>.
+     * Getter for <code>transactions.is_duplicate</code>.
      */
-    public Boolean getIsDuplicate() {
+    public Integer getIsDuplicate() {
         return this.isDuplicate;
     }
 
     /**
-     * Setter for <code>minance.transactions.is_duplicate</code>.
+     * Setter for <code>transactions.is_duplicate</code>.
      */
-    public void setIsDuplicate(Boolean isDuplicate) {
+    public void setIsDuplicate(Integer isDuplicate) {
         this.isDuplicate = isDuplicate;
     }
 
