@@ -6,10 +6,7 @@ package today.ihelio.jooq.tables.records;
 
 import java.math.BigDecimal;
 
-import org.jooq.Field;
 import org.jooq.Record1;
-import org.jooq.Record6;
-import org.jooq.Row6;
 import org.jooq.impl.UpdatableRecordImpl;
 
 import today.ihelio.jooq.tables.Accounts;
@@ -18,90 +15,90 @@ import today.ihelio.jooq.tables.Accounts;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class AccountsRecord extends UpdatableRecordImpl<AccountsRecord> implements Record6<Integer, Integer, String, String, String, BigDecimal> {
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class AccountsRecord extends UpdatableRecordImpl<AccountsRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * Setter for <code>minance.accounts.account_id</code>.
+     * Setter for <code>accounts.account_id</code>.
      */
     public void setAccountId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.account_id</code>.
+     * Getter for <code>accounts.account_id</code>.
      */
     public Integer getAccountId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for <code>minance.accounts.bank_id</code>.
+     * Setter for <code>accounts.bank_id</code>.
      */
     public void setBankId(Integer value) {
         set(1, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.bank_id</code>.
+     * Getter for <code>accounts.bank_id</code>.
      */
     public Integer getBankId() {
         return (Integer) get(1);
     }
 
     /**
-     * Setter for <code>minance.accounts.bank_name</code>.
+     * Setter for <code>accounts.bank_name</code>.
      */
     public void setBankName(String value) {
         set(2, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.bank_name</code>.
+     * Getter for <code>accounts.bank_name</code>.
      */
     public String getBankName() {
         return (String) get(2);
     }
 
     /**
-     * Setter for <code>minance.accounts.account_name</code>.
+     * Setter for <code>accounts.account_name</code>.
      */
     public void setAccountName(String value) {
         set(3, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.account_name</code>.
+     * Getter for <code>accounts.account_name</code>.
      */
     public String getAccountName() {
         return (String) get(3);
     }
 
     /**
-     * Setter for <code>minance.accounts.account_type</code>.
+     * Setter for <code>accounts.account_type</code>.
      */
     public void setAccountType(String value) {
         set(4, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.account_type</code>.
+     * Getter for <code>accounts.account_type</code>.
      */
     public String getAccountType() {
         return (String) get(4);
     }
 
     /**
-     * Setter for <code>minance.accounts.init_balance</code>.
+     * Setter for <code>accounts.init_balance</code>.
      */
     public void setInitBalance(BigDecimal value) {
         set(5, value);
     }
 
     /**
-     * Getter for <code>minance.accounts.init_balance</code>.
+     * Getter for <code>accounts.init_balance</code>.
      */
     public BigDecimal getInitBalance() {
         return (BigDecimal) get(5);
@@ -114,157 +111,6 @@ public class AccountsRecord extends UpdatableRecordImpl<AccountsRecord> implemen
     @Override
     public Record1<Integer> key() {
         return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record6 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row6<Integer, Integer, String, String, String, BigDecimal> fieldsRow() {
-        return (Row6) super.fieldsRow();
-    }
-
-    @Override
-    public Row6<Integer, Integer, String, String, String, BigDecimal> valuesRow() {
-        return (Row6) super.valuesRow();
-    }
-
-    @Override
-    public Field<Integer> field1() {
-        return Accounts.ACCOUNTS.ACCOUNT_ID;
-    }
-
-    @Override
-    public Field<Integer> field2() {
-        return Accounts.ACCOUNTS.BANK_ID;
-    }
-
-    @Override
-    public Field<String> field3() {
-        return Accounts.ACCOUNTS.BANK_NAME;
-    }
-
-    @Override
-    public Field<String> field4() {
-        return Accounts.ACCOUNTS.ACCOUNT_NAME;
-    }
-
-    @Override
-    public Field<String> field5() {
-        return Accounts.ACCOUNTS.ACCOUNT_TYPE;
-    }
-
-    @Override
-    public Field<BigDecimal> field6() {
-        return Accounts.ACCOUNTS.INIT_BALANCE;
-    }
-
-    @Override
-    public Integer component1() {
-        return getAccountId();
-    }
-
-    @Override
-    public Integer component2() {
-        return getBankId();
-    }
-
-    @Override
-    public String component3() {
-        return getBankName();
-    }
-
-    @Override
-    public String component4() {
-        return getAccountName();
-    }
-
-    @Override
-    public String component5() {
-        return getAccountType();
-    }
-
-    @Override
-    public BigDecimal component6() {
-        return getInitBalance();
-    }
-
-    @Override
-    public Integer value1() {
-        return getAccountId();
-    }
-
-    @Override
-    public Integer value2() {
-        return getBankId();
-    }
-
-    @Override
-    public String value3() {
-        return getBankName();
-    }
-
-    @Override
-    public String value4() {
-        return getAccountName();
-    }
-
-    @Override
-    public String value5() {
-        return getAccountType();
-    }
-
-    @Override
-    public BigDecimal value6() {
-        return getInitBalance();
-    }
-
-    @Override
-    public AccountsRecord value1(Integer value) {
-        setAccountId(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord value2(Integer value) {
-        setBankId(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord value3(String value) {
-        setBankName(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord value4(String value) {
-        setAccountName(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord value5(String value) {
-        setAccountType(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord value6(BigDecimal value) {
-        setInitBalance(value);
-        return this;
-    }
-
-    @Override
-    public AccountsRecord values(Integer value1, Integer value2, String value3, String value4, String value5, BigDecimal value6) {
-        value1(value1);
-        value2(value2);
-        value3(value3);
-        value4(value4);
-        value5(value5);
-        value6(value6);
-        return this;
     }
 
     // -------------------------------------------------------------------------
