@@ -4,10 +4,7 @@
 package today.ihelio.jooq.tables.records;
 
 
-import org.jooq.Field;
 import org.jooq.Record1;
-import org.jooq.Record2;
-import org.jooq.Row2;
 import org.jooq.impl.UpdatableRecordImpl;
 
 import today.ihelio.jooq.tables.MinanceCategory;
@@ -16,34 +13,34 @@ import today.ihelio.jooq.tables.MinanceCategory;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class MinanceCategoryRecord extends UpdatableRecordImpl<MinanceCategoryRecord> implements Record2<Integer, String> {
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class MinanceCategoryRecord extends UpdatableRecordImpl<MinanceCategoryRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * Setter for <code>minance.minance_category.m_category_id</code>.
+     * Setter for <code>minance_category.m_category_id</code>.
      */
     public void setMCategoryId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for <code>minance.minance_category.m_category_id</code>.
+     * Getter for <code>minance_category.m_category_id</code>.
      */
     public Integer getMCategoryId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for <code>minance.minance_category.category</code>.
+     * Setter for <code>minance_category.category</code>.
      */
     public void setCategory(String value) {
         set(1, value);
     }
 
     /**
-     * Getter for <code>minance.minance_category.category</code>.
+     * Getter for <code>minance_category.category</code>.
      */
     public String getCategory() {
         return (String) get(1);
@@ -56,69 +53,6 @@ public class MinanceCategoryRecord extends UpdatableRecordImpl<MinanceCategoryRe
     @Override
     public Record1<Integer> key() {
         return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record2 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row2<Integer, String> fieldsRow() {
-        return (Row2) super.fieldsRow();
-    }
-
-    @Override
-    public Row2<Integer, String> valuesRow() {
-        return (Row2) super.valuesRow();
-    }
-
-    @Override
-    public Field<Integer> field1() {
-        return MinanceCategory.MINANCE_CATEGORY.M_CATEGORY_ID;
-    }
-
-    @Override
-    public Field<String> field2() {
-        return MinanceCategory.MINANCE_CATEGORY.CATEGORY;
-    }
-
-    @Override
-    public Integer component1() {
-        return getMCategoryId();
-    }
-
-    @Override
-    public String component2() {
-        return getCategory();
-    }
-
-    @Override
-    public Integer value1() {
-        return getMCategoryId();
-    }
-
-    @Override
-    public String value2() {
-        return getCategory();
-    }
-
-    @Override
-    public MinanceCategoryRecord value1(Integer value) {
-        setMCategoryId(value);
-        return this;
-    }
-
-    @Override
-    public MinanceCategoryRecord value2(String value) {
-        setCategory(value);
-        return this;
-    }
-
-    @Override
-    public MinanceCategoryRecord values(Integer value1, String value2) {
-        value1(value1);
-        value2(value2);
-        return this;
     }
 
     // -------------------------------------------------------------------------
