@@ -4,10 +4,7 @@
 package today.ihelio.jooq.tables.records;
 
 
-import org.jooq.Field;
 import org.jooq.Record1;
-import org.jooq.Record3;
-import org.jooq.Row3;
 import org.jooq.impl.UpdatableRecordImpl;
 
 import today.ihelio.jooq.tables.RawCategoryToMinanceCategory;
@@ -16,38 +13,34 @@ import today.ihelio.jooq.tables.RawCategoryToMinanceCategory;
 /**
  * This class is generated by jOOQ.
  */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class RawCategoryToMinanceCategoryRecord extends UpdatableRecordImpl<RawCategoryToMinanceCategoryRecord> implements Record3<Integer, String, Integer> {
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class RawCategoryToMinanceCategoryRecord extends UpdatableRecordImpl<RawCategoryToMinanceCategoryRecord> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * Setter for
-     * <code>minance.raw_category_to_minance_category.rc_to_mc_id</code>.
+     * Setter for <code>raw_category_to_minance_category.rc_to_mc_id</code>.
      */
     public void setRcToMcId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for
-     * <code>minance.raw_category_to_minance_category.rc_to_mc_id</code>.
+     * Getter for <code>raw_category_to_minance_category.rc_to_mc_id</code>.
      */
     public Integer getRcToMcId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for
-     * <code>minance.raw_category_to_minance_category.raw_category</code>.
+     * Setter for <code>raw_category_to_minance_category.raw_category</code>.
      */
     public void setRawCategory(String value) {
         set(1, value);
     }
 
     /**
-     * Getter for
-     * <code>minance.raw_category_to_minance_category.raw_category</code>.
+     * Getter for <code>raw_category_to_minance_category.raw_category</code>.
      */
     public String getRawCategory() {
         return (String) get(1);
@@ -55,7 +48,7 @@ public class RawCategoryToMinanceCategoryRecord extends UpdatableRecordImpl<RawC
 
     /**
      * Setter for
-     * <code>minance.raw_category_to_minance_category.minance_category_id</code>.
+     * <code>raw_category_to_minance_category.minance_category_id</code>.
      */
     public void setMinanceCategoryId(Integer value) {
         set(2, value);
@@ -63,7 +56,7 @@ public class RawCategoryToMinanceCategoryRecord extends UpdatableRecordImpl<RawC
 
     /**
      * Getter for
-     * <code>minance.raw_category_to_minance_category.minance_category_id</code>.
+     * <code>raw_category_to_minance_category.minance_category_id</code>.
      */
     public Integer getMinanceCategoryId() {
         return (Integer) get(2);
@@ -76,91 +69,6 @@ public class RawCategoryToMinanceCategoryRecord extends UpdatableRecordImpl<RawC
     @Override
     public Record1<Integer> key() {
         return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record3 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row3<Integer, String, Integer> fieldsRow() {
-        return (Row3) super.fieldsRow();
-    }
-
-    @Override
-    public Row3<Integer, String, Integer> valuesRow() {
-        return (Row3) super.valuesRow();
-    }
-
-    @Override
-    public Field<Integer> field1() {
-        return RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY.RC_TO_MC_ID;
-    }
-
-    @Override
-    public Field<String> field2() {
-        return RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY.RAW_CATEGORY;
-    }
-
-    @Override
-    public Field<Integer> field3() {
-        return RawCategoryToMinanceCategory.RAW_CATEGORY_TO_MINANCE_CATEGORY.MINANCE_CATEGORY_ID;
-    }
-
-    @Override
-    public Integer component1() {
-        return getRcToMcId();
-    }
-
-    @Override
-    public String component2() {
-        return getRawCategory();
-    }
-
-    @Override
-    public Integer component3() {
-        return getMinanceCategoryId();
-    }
-
-    @Override
-    public Integer value1() {
-        return getRcToMcId();
-    }
-
-    @Override
-    public String value2() {
-        return getRawCategory();
-    }
-
-    @Override
-    public Integer value3() {
-        return getMinanceCategoryId();
-    }
-
-    @Override
-    public RawCategoryToMinanceCategoryRecord value1(Integer value) {
-        setRcToMcId(value);
-        return this;
-    }
-
-    @Override
-    public RawCategoryToMinanceCategoryRecord value2(String value) {
-        setRawCategory(value);
-        return this;
-    }
-
-    @Override
-    public RawCategoryToMinanceCategoryRecord value3(Integer value) {
-        setMinanceCategoryId(value);
-        return this;
-    }
-
-    @Override
-    public RawCategoryToMinanceCategoryRecord values(Integer value1, String value2, Integer value3) {
-        value1(value1);
-        value2(value2);
-        value3(value3);
-        return this;
     }
 
     // -------------------------------------------------------------------------
