@@ -13,43 +13,46 @@ import static today.ihelio.minance.csvpojos.BankAccountPair.BankName.CITI;
 
 @QuarkusTest
 public class BankServiceTest {
-  @Inject BankService bankService;
-  @Inject Flyway flyway;
+	@Inject
+	BankService bankService;
+	@Inject
+	Flyway flyway;
 
-  @BeforeEach
-  void setUp() {
-    flyway.migrate();
-  }
+	@BeforeEach
+	void setUp() {
+		flyway.clean();
+		flyway.migrate();
+	}
 
-  @AfterEach
-  void clean() {
-    flyway.clean();
-  }
+	@AfterEach
+	void clean() {
+		flyway.clean();
+	}
 
-  @Test
-  public void testBankService_CreateDeleteBanks() throws Exception {
-    assertThat(bankService.create(CITI)).isEqualTo(1);
+	@Test
+	public void testBankService_CreateDeleteBanks() throws Exception {
+		assertThat(bankService.create(CITI)).isEqualTo(1);
 
-    assertThat(bankService.retrieveAll().size()).isEqualTo(1);
+		assertThat(bankService.retrieveAll().size()).isEqualTo(1);
 
-    assertThat(bankService.delete(CITI)).isEqualTo(1);
+		assertThat(bankService.delete(CITI)).isEqualTo(1);
 
-    assertThat(bankService.retrieveAll().size()).isEqualTo(0);
-  }
+		assertThat(bankService.retrieveAll().size()).isEqualTo(0);
+	}
 
-  @Test
-  public void testBankService_RetrieveBanks() throws Exception {
-    assertThat(bankService.create(CITI)).isEqualTo(1);
+	@Test
+	public void testBankService_RetrieveBanks() throws Exception {
+		assertThat(bankService.create(CITI)).isEqualTo(1);
 
-    Banks citi = bankService.findBankByName(CITI);
-    assertThat(citi.getBankName()).isEqualTo(CITI.getName());
+		Banks citi = bankService.findBankByName(CITI);
+		assertThat(citi.getBankName()).isEqualTo(CITI.getName());
 
-    assertThat(bankService.retrieve(citi.getBankId())).isEqualTo(citi);
-  }
+		assertThat(bankService.retrieve(citi.getBankId())).isEqualTo(citi);
+	}
 
-  @Test
-  public void testBankService_RetrieveNonexistentBanks() throws Exception {
-    Banks citi = bankService.findBankByName(CITI);
-    assertThat(citi).isNull();
-  }
+	@Test
+	public void testBankService_RetrieveNonexistentBanks() throws Exception {
+		Banks citi = bankService.findBankByName(CITI);
+		assertThat(citi).isNull();
+	}
 }
